requirements.txt
fastapi
uvicorn
pyzmq
websockets
opencv-python
numpy
cryptography
